add pubsub for control map
