+++ editor-server/src/resolvers/subscriptions/controlMap.ts
@@ -1,11 +1,17 @@
 import {ObjectType, Field, ID} from "type-graphql"
+import { ObjectId } from "mongodb";
+import { ControlMapScalar } from "../types/controlMap";
 
 @ObjectType()
 export class ControlMapPayload {
     @Field()
     mutation: ControlMapMutation;
 
+    @Field(type => ControlMapScalar)
+    frames: any[]   
     
+    @Field()
+    editBy: string; 
 }
 
 export enum ControlMapMutation {
